Improving Email DRYness (#2486)
* Email DRY

* WIP

* Improve email DRYness
