+++ apps/web/lib/emails/templates/organizer-scheduled-email.ts
@@ -4,21 +4,20 @@ import timezone from "dayjs/plugin/timezone";
 import toArray from "dayjs/plugin/toArray";
 import utc from "dayjs/plugin/utc";
 import { createEvent, DateArray, Person } from "ics";
-import nodemailer from "nodemailer";
 import rrule from "rrule";
 
 import { getAppName } from "@calcom/app-store/utils";
 import { getCancelLink, getRichDescription } from "@calcom/lib/CalEventParser";
-import { getErrorFromUnknown } from "@calcom/lib/errors";
-import { serverConfig } from "@calcom/lib/serverConfig";
 import type { CalendarEvent, RecurringEvent } from "@calcom/types/Calendar";
 
+import BaseEmail from "@lib/emails/templates/_base-email";
+
 import {
-  emailHead,
-  emailSchedulingBodyHeader,
   emailBodyLogo,
+  emailHead,
   emailScheduledBodyHeaderContent,
   emailSchedulingBodyDivider,
+  emailSchedulingBodyHeader,
   linkIcon,
 } from "./common";
 
@@ -27,30 +26,15 @@ dayjs.extend(timezone);
 dayjs.extend(localizedFormat);
 dayjs.extend(toArray);
 
-export default class OrganizerScheduledEmail {
+export default class OrganizerScheduledEmail extends BaseEmail {
   calEvent: CalendarEvent;
   recurringEvent: RecurringEvent;
 
   constructor(calEvent: CalendarEvent, recurringEvent: RecurringEvent) {
+    super();
+    this.name = "SEND_BOOKING_CONFIRMATION";
     this.calEvent = calEvent;
     this.recurringEvent = recurringEvent;
-  }
-
-  public sendEmail() {
-    new Promise((resolve, reject) =>
-      nodemailer
-        .createTransport(this.getMailerOptions().transport)
-        .sendMail(this.getNodeMailerPayload(), (_err, info) => {
-          if (_err) {
-            const err = getErrorFromUnknown(_err);
-            this.printNodeMailerError(err);
-            reject(err);
-          } else {
-            resolve(info);
-          }
-        })
-    ).catch((e) => console.error("sendEmail", e));
-    return new Promise((resolve) => resolve("send mail async"));
   }
 
   protected getiCalEventAsString(): string | undefined {
@@ -121,13 +105,6 @@ export default class OrganizerScheduledEmail {
     };
   }
 
-  protected getMailerOptions() {
-    return {
-      transport: serverConfig.transport,
-      from: serverConfig.from,
-    };
-  }
-
   protected getTextBody(): string {
     return `
 ${this.calEvent.organizer.language.translate(
@@ -137,10 +114,6 @@ ${this.calEvent.organizer.language.translate("emailed_you_and_any_other_attendee
 
 ${getRichDescription(this.calEvent)}
 `.trim();
-  }
-
-  protected printNodeMailerError(error: Error): void {
-    console.error("SEND_BOOKING_CONFIRMATION_ERROR", this.calEvent.organizer.email, error);
   }
 
   protected getHtmlBody(): string {
